arm64: dts: qcom: sc8280xp-sp9-5G: Add pmic regulator supplies
The s11b, s12b and s1c supplies are used by several pmic regulators.
Add the missing description to the Microsoft Surface 9 Pro 5G devicetree.

Note that there are still some consumers that are not (fully) described
in the devicetree so the supplies must remain marked as always-on for
now.

Signed-off-by: Jérôme de Bretagne <jerome.debretagne@gmail.com>